move metering-report test
